requirements.txt
fastapi
fastapi-utilities
requests
python-dotenv
uvicorn
websockets